+++ common/src/main/scala/dx/util/Paths.scala
@@ -1,59 +1,65 @@
 package dx.util
 
-import java.nio.file.{Files, Path}
-
+import java.nio.file.Files
 import spray.json.{JsString, JsValue}
 
 trait EvalPaths {
-  def getRootDir(ensureExists: Boolean = false): Path
+  def getRootDir(ensureExists: Boolean = false): PosixPath
 
-  def getTempDir(ensureExists: Boolean = false): Path
+  def getTempDir(ensureExists: Boolean = false): PosixPath
 
   /**
     * The execution directory - used as the base dir for relative paths (e.g. for glob search).
     */
-  def getWorkDir(ensureExists: Boolean = false): Path
+  def getWorkDir(ensureExists: Boolean = false): PosixPath
 
-  def getMetaDir(ensureExists: Boolean = false): Path
+  def getMetaDir(ensureExists: Boolean = false): PosixPath
 
   /**
     * The file that has a copy of standard output.
     */
-  def getStdoutFile(ensureParentExists: Boolean = false): Path
+  def getStdoutFile(ensureParentExists: Boolean = false): PosixPath
 
   /**
     * The file that has a copy of standard error.
     */
-  def getStderrFile(ensureParentExists: Boolean = false): Path
+  def getStderrFile(ensureParentExists: Boolean = false): PosixPath
 }
 
 abstract class BaseEvalPaths extends EvalPaths {
-  private var cache: Map[String, Path] = Map.empty
+  private var cache: Map[String, PosixPath] = Map.empty
 
-  protected def createDir(key: String, path: Path): Path = {
-    val resolved = FileUtils.createDirectories(path)
+  protected def createDir(key: String, path: PosixPath): PosixPath = {
+    val resolved = PosixPath(FileUtils.createDirectories(path.asJavaPath).toString)
     cache += (key -> resolved)
     resolved
   }
 
-  protected def getOrCreateDir(key: String, path: Path, ensureExists: Boolean): Path = {
-    val resolved = cache.getOrElse(key, if (ensureExists) createDir(key, path) else path)
-    if (Files.exists(resolved)) {
-      resolved.toRealPath()
-    } else {
-      resolved.toAbsolutePath
-    }
+  protected def getOrCreateDir(key: String, path: PosixPath, ensureExists: Boolean): PosixPath = {
+    cache.getOrElse(
+        key,
+        if (ensureExists) {
+          createDir(key, path)
+        } else {
+          val localPath = path.asJavaPath
+          if (Files.exists(localPath)) {
+            PosixPath(localPath.toRealPath().toString)
+          } else {
+            path
+          }
+        }
+    )
   }
 }
 
 trait ExecPaths extends EvalPaths {
-  def getCommandFile(ensureParentExists: Boolean = false): Path
+  def getCommandFile(ensureParentExists: Boolean = false): PosixPath
 
-  def getReturnCodeFile(ensureParentExists: Boolean = false): Path
+  def getReturnCodeFile(ensureParentExists: Boolean = false): PosixPath
 
-  def getContainerCommandFile(ensureParentExists: Boolean = false): Path
+  def getContainerCommandFile(ensureParentExists: Boolean = false): PosixPath
 
-  def getContainerIdFile(ensureParentExists: Boolean = false): Path
+  def getContainerIdFile(ensureParentExists: Boolean = false): PosixPath
 
   def toJson(onlyExisting: Boolean = true): Map[String, JsValue] = {
     Map(
@@ -69,7 +75,7 @@ trait ExecPaths extends EvalPaths {
         "containerId" -> getContainerIdFile()
     ).flatMap {
       case (key, path) =>
-        if (!onlyExisting || Files.exists(path)) {
+        if (!onlyExisting || Files.exists(path.asJavaPath)) {
           Some(key -> JsString(path.toString))
         } else {
           None