+++ common/RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # dxCommon
 
+## in develop
+
+* Adds `PosixPath` class for working with POSIX-style paths
+* **Breaking**: changes `Paths` to use `PosixPath` rather than `java.nio.Path`
+
 ## 0.9.0 (2021-11-15)
 
 * Adds option to `Logger.trace*` to show beginning and/or end of log when limiting trace length